Variant: Add support for binary payload

+++ platform/gucefCORE/src/gucefCORE_CUtf8String.cpp
@@ -710,7 +710,7 @@ CUtf8String::Reserve( const UInt32 byteSize, Int32 newLength )
 
         if ( byteSize < m_byteSize )
             m_length = 0;
-        if ( newLength >= 0 && newLength < byteSize )
+        if ( newLength >= 0 && (UInt32) newLength < byteSize )
             m_length = (UInt32) newLength;            
 
         m_byteSize = byteSize;
